backend/services/alerts_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
minio==7.2.0
reportlab==4.0.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1